+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## 0.2.3
  - Added last_request, last_response to Eloqua::API
 
+ - Added wrapper around the Savon SOAP and HTTP exceptions.
+   Savon should disable its error handling and let Eloqua raise its own
+   errors by using `Savon::configure {|config| config.raise_errors =
+   false }`. This allows us to capture last_request and last_response
+   even when there is a Soap or HTTP fault
+
 ## 0.2
  - Abstracted out majority of requests into Eloqua::Api::Service